Add skills informations

+++ src/main/java/fr/arko/dokbuild/dao/CardsDao.java
@@ -11,7 +11,7 @@ import org.springframework.data.repository.query.Param;
 import fr.arko.dokbuild.domain.Cards;
 
 
-public interface CardsDao extends JpaRepository<Cards, Long> {
+public interface CardsDao extends JpaRepository<Cards, Integer> {
 	
 	@Query("SELECT c FROM Cards c INNER JOIN Cards c2 ON c.awakedCardId=c2.id WHERE c.name like %:name% and c.rarity in :rarity and c2.element in :element and c.cost > 1")
 	Page<Cards> searchCard(@Param("name") String name, @Param("rarity") List<Integer> rarity, @Param("element") List<Integer> element, Pageable pageable);
