+++ src/main/java/fr/arko/dokbuild/dao/PassiveSkillsDao.java
@@ -2,10 +2,9 @@ package fr.arko.dokbuild.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import fr.arko.dokbuild.domain.Cards;
 import fr.arko.dokbuild.domain.PassiveSkills;
 
 
-public interface PassiveSkillsDao extends JpaRepository<PassiveSkills, Long> {
+public interface PassiveSkillsDao extends JpaRepository<PassiveSkills, Integer> {
 	
 }
